Modified CKEditor to normal usage
Previously, I was using the jQuery instance of CKEditor and when I
needed to use back to the default functions of the editor, I can't.

I will need the control over the editor as much as possible because
every time I set the editor's content, the cursor moves to the start.
With this access to other functions, I can easily set where the cursor
belongs.